services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
redis==5.1.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
eth-account>=0.11.0
web3>=6.15.0
h3==3.7.7
httpx==0.27.0
python-multipart==0.0.9
pytest==8.3.3
pytest-asyncio==0.24.0
